UI/Service - implemeneted Sales Overview

+++ aspnet-core/src/caikdduopos.Domain/Entity/Sales.cs
@@ -5,13 +5,13 @@ namespace caikdduopos.Entity
 {
     public class Sales:Entity<Guid>
     {
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public Member Member { get; set; }
         public SalesType SalesType { get; set; }
         public PaymentMethod PaymentMethod { get; set; }
         public Product Product { get; set; }
         public int Quantity { get; set; }
-        public decimal Discount { get; set; }
+        public double Discount { get; set; }
        
         public int SalesTypeId { get; set; }
         public int PaymentMethodId { get; set; }
